Separate pipes and providers
